+++ crates/sui-benchmark/Cargo.toml
@@ -39,7 +39,7 @@ sui-core = { path = "../sui-core" }
 sui-config = { path = "../sui-config" }
 sui-network = { path = "../sui-network" }
 sui-types = { path = "../sui-types" }
-sui-sdk = { path = "../sui-sdk" }
+sui-keys = { path = "../sui-keys" }
 sui-node = { path = "../sui-node" }
 sui-json-rpc-types = { path = "../sui-json-rpc-types" }
