+++ crates/sui-keys/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sui-keys"
+version = "0.0.0"
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.64"
+serde = { version = "1.0.144", features = ["derive"] }
+serde_json = "1.0.83"
+signature = "1.6.0"
+rand = "0.8.5"
+tiny-bip39 = "1.0.0"
+bip32 = "0.4.0"
+slip10_ed25519 = "0.1.3"
+fastcrypto = { workspace = true, features = ["copy_key"] }
+
+sui-types = { path = "../sui-types" }
+
+workspace-hack.workspace = true
+
+[dev-dependencies]
+tempfile = "3.3.0"
+sha3 = "0.10.4"
