+++ crates/sui-keys/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod key_derive;
+pub mod keystore;
