+++ crates/sui-types/Cargo.toml
@@ -38,8 +38,6 @@ strum_macros = "^0.24"
 roaring = "0.10.1"
 enum_dispatch = "^0.3"
 eyre = "0.6.8"
-bip32 = "0.4.0"
-slip10_ed25519 = "0.1.3"
 
 name-variant.workspace = true
 typed-store.workspace = true
