+++ crates/sui/Cargo.toml
@@ -29,6 +29,7 @@ sui-json = { path = "../sui-json" }
 sui-swarm = { path = "../sui-swarm" }
 sui-json-rpc-types = { path = "../sui-json-rpc-types" }
 sui-sdk = { path = "../sui-sdk" }
+sui-keys = { path = "../sui-keys" }
 
 fastcrypto.workspace = true
