+++ crates/sui/src/unit_tests/keytool_tests.rs
@@ -8,10 +8,7 @@ use super::write_keypair_to_file;
 use super::KeyToolCommand;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
-use sui_sdk::crypto::AccountKeystore;
-use sui_sdk::crypto::FileBasedKeystore;
-use sui_sdk::crypto::InMemKeystore;
-use sui_sdk::crypto::Keystore;
+use sui_keys::keystore::{AccountKeystore, FileBasedKeystore, InMemKeystore, Keystore};
 use sui_types::base_types::SuiAddress;
 use sui_types::crypto::get_key_pair;
 use sui_types::crypto::get_key_pair_from_rng;
