+++ crates/sui/tests/checkpoints_tests.rs
@@ -10,9 +10,9 @@ use sui_core::{
     authority_aggregator::{AuthorityAggregator, AuthorityAggregatorBuilder},
     authority_client::NetworkAuthorityClient,
 };
+use sui_keys::keystore::{InMemKeystore, Keystore};
 use sui_macros::sim_test;
 use sui_node::SuiNodeHandle;
-use sui_sdk::crypto::{InMemKeystore, Keystore};
 use sui_types::{
     base_types::{ExecutionDigests, TransactionDigest},
     messages::{
