+++ crates/test-utils/Cargo.toml
@@ -29,6 +29,7 @@ sui-json-rpc-types= { path = "../sui-json-rpc-types" }
 sui-node = { path = "../sui-node" }
 sui-swarm = { path = "../sui-swarm" }
 sui-types = { path = "../sui-types" }
+sui-keys = { path = "../sui-keys" }
 sui-sdk = { path = "../sui-sdk" }
 once_cell = "1.14.0"
 multiaddr = "0.14.0"
