+++ crates/test-utils/src/messages.rs
@@ -12,8 +12,8 @@ use sui::client_commands::{SuiClientCommandResult, SuiClientCommands};
 use sui_adapter::genesis;
 use sui_core::test_utils::to_sender_signed_transaction;
 use sui_json_rpc_types::SuiObjectInfo;
-use sui_sdk::crypto::AccountKeystore;
-use sui_sdk::crypto::Keystore;
+use sui_keys::keystore::AccountKeystore;
+use sui_keys::keystore::Keystore;
 use sui_types::base_types::ObjectRef;
 use sui_types::base_types::SuiAddress;
 use sui_types::base_types::{ObjectDigest, ObjectID, SequenceNumber};
